Remove validators from Controllers

+++ src/modules/customers/services/RemoveCustomerService.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 import ICustomersRepository from '../repositories/ICustomersRepository';
 import Customer from '../entities/Customer';
+import AppError from '@shared/errors/AppError';
 
 export interface IRemoveCustomerData {
   _id: string;
@@ -18,6 +19,8 @@ export default class RemoveCustomerService {
   public async execute({ _id }: IRemoveCustomerData): Promise<void> {
     const customerToRemove = new Customer;
     customerToRemove._id = _id;
+
+    if (!_id) throw new AppError('Customer id is required', 400);
 
     await this.customersRepository.remove(customerToRemove);
   }
